+++ satch.h
@@ -62,7 +62,7 @@ int satch_val (struct satch *, int literal);
 // Short hands for adding clauses.
 
 void satch_add_empty (struct satch *);
-void satch_add_unit (struct satch *, int );
+void satch_add_unit (struct satch *, int);
 void satch_add_binary_clause (struct satch *, int, int);
 void satch_add_ternary_clause (struct satch *, int, int, int);
 void satch_add_quaternary_clause (struct satch *, int, int, int, int);